Add app_configuration decorator and shopify_api initializer to config/initializers
